third_party/blink/web_tests/shadow-dom/crashes/slots-in-v0-crash.html

<!DOCTYPE html>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<div id=host></div>
<script>
test(() => {
  let shadowRoot = host.attachShadow({mode: 'open'});
  let slot1 = document.createElement('slot');
  shadowRoot.appendChild(slot1);
  document.body.appendChild(slot1);
  let slot2 = document.createElement('slot');
  shadowRoot.appendChild(slot2);
  assert_true(true, 'This test must run without crash.');
}, 'Adding/removing slots in a v0 shadow tree should not cause a crash (crbug.com/634506).' );
</script>
